Update README.md
This seemed like a fun and obtainable challenge to compete in.

Our web application has an authentication system for users to login, and It can easily be compiled to iOS and android to run as a native mobile app.

Our interface has a points system to help make it fun to absorb the information. Points can be awarded by the user scrolling through reading sections, answering questions about the reading questions, and by winning mini games.

Microbi was built using node.js for the web server, jQuery as the primary JavaScript library to build the interface, and HTML5 and CSS. D3JS, a popular graphing library was also used for our Microbiology tree chart.

The largest challenges we ran into were design decisions about how the user interface should look, which features to implement that would make the app the most interesting, and not having enough time to implement a few more features.

We're proud of the reading interface, and a lot of small scripts that do clever things.

we learned how to work with a node.js templating engine, ejs. The minor syntax changes take a little getting used to.

What is next for Microbi? We would like to take it live. With more content and features implemented at a production quality level, this application could go live. Servers cost money, and we think this app can produce more revenue than the cost of the server.